+++ InfiniteFibonacci/IFFiboFinder.m
@@ -10,7 +10,40 @@
 
 @implementation IFFiboFinder
 
--(NSUInteger)findFiboNumber:(NSInteger)someNumber
+//-(NSUInteger)findFiboNumber:(NSInteger)someNumber
+//{
+//    NSUInteger isBaseCase = [self checkFiboBaseCaseWithNumber:someNumber];
+//    
+//    if (isBaseCase != -1) {
+//        return isBaseCase;
+//    } else {
+//        return [self calculateFiboNumber:someNumber];
+//    }
+//}
+//
+//- (NSUInteger)checkFiboBaseCaseWithNumber:(NSInteger)baseCaseNumber
+//{
+//    if (baseCaseNumber < 2) {
+//        return baseCaseNumber;
+//    } else {
+//        return -1;
+//    }
+//}
+//
+//- (NSUInteger)calculateFiboNumber:(NSInteger)fiboCalculation
+//{
+//    NSInteger twoFiboAgo = 0;
+//    NSInteger oneFiboAgo = 1;
+//    NSInteger currentFiboNumber = 0;
+//    for (NSInteger i = 2; i <= fiboCalculation; i++) {
+//        currentFiboNumber = twoFiboAgo + oneFiboAgo;
+//        twoFiboAgo = oneFiboAgo;
+//        oneFiboAgo = currentFiboNumber;
+//    }
+//    return currentFiboNumber;
+//}
+
+-(unsigned long long)findFiboNumber:(NSInteger)someNumber
 {
     NSUInteger isBaseCase = [self checkFiboBaseCaseWithNumber:someNumber];
